# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ellowring API
  description: |
    Ellowring Software Solutions — Enterprise SaaS REST API (Phase 5).
    Learn. Prepare. Build. Get Hired.
  version: 1.0.0
  contact:
    name: Ellowring Engineering
    email: engineering@ellowring.com
servers:
  - url: http://localhost:4000/api
    description: Local
  - url: https://api-staging.ellowring.com/api
    description: Staging
  - url: https://api.ellowring.com/api
    description: Production
...
paths:
...
  /v1/students/me/dashboard:
    get:
      tags:
        - Student
      summary: "Student home KPIs: enrollments, applications, wallet, notifications."
      operationId: get__students_me_dashboard
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/students/me:
    get:
      tags:
        - Student
      summary: Get student profile extension.
      operationId: get__students_me
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    patch:
      tags:
        - Student
      summary: "Update student profile (grade, stream, city, bio, resume)."
      operationId: patch__students_me
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/students/me/courses:
    get:
      tags:
        - Student
      summary: Enrolled courses + progress.
      operationId: get__students_me_courses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/students/me/certificates:
    get:
      tags:
        - Student
      summary: Student certificates.
      operationId: get__students_me_certificates
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/students/me/bookmarks:
    get:
      tags:
        - Student
      summary: Bookmarks list.
      operationId: get__students_me_bookmarks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    post:
      tags:
        - Student
      summary: "Create bookmark {entityType, entityId}."
      operationId: post__students_me_bookmarks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/students/me/bookmarks/{id}:
    delete:
      tags:
        - Student
      summary: Remove bookmark.
      operationId: delete__students_me_bookmarks__id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/students/me/favorites:
    get:
      tags:
        - Student
      summary: Favorites list.
      operationId: get__students_me_favorites
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    post:
      tags:
        - Student
      summary: Add favorite.
      operationId: post__students_me_favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/students/me/notifications:
    get:
      tags:
        - Student
      summary: Student notifications inbox.
      operationId: get__students_me_notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data: {}
        meta:
          type: object
          properties:
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrev: { type: boolean }